+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/AlbumEditPanel.java
@@ -19,7 +19,7 @@ public final class AlbumEditPanel extends EntityEditPanel {
 
   public AlbumEditPanel(SwingEntityEditModel editModel) {
     super(editModel);
-    setDefaultTextFieldColumns(15);
+    defaultTextFieldColumns().set(15);
   }
 
   @Override