+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/CustomerEditPanel.java
@@ -21,7 +21,7 @@ public final class CustomerEditPanel extends EntityEditPanel {
 
   public CustomerEditPanel(SwingEntityEditModel editModel) {
     super(editModel);
-    setDefaultTextFieldColumns(12);
+    defaultTextFieldColumns().set(12);
   }
 
   @Override